Change to modified Google formatting style.
Major changes from old style:
- 80 column limit
- 2-space indent

The main change from Google style is that the { at the beginning of a function gets its own line. Reformatting was mostly done with Clang-format, but some hand editing was necessary.

There are further style changes I want to make, but at the moment neither Uncrustify nor Clang-format is up to the task.